Fix iOS 7.x+64 bits Crash, the following code is the responsible for crash
NSIndexPath* index = [NSIndexPath indexPathForRow:row inSection:section];
    objc_setAssociatedObject(index, &kColumnObjectKey, @(column), OBJC_ASSOCIATION_RETAIN_NONATOMIC);

The problem is that this NSIndexPath is a tagged pointer (i.e. the data is kept entirely within the pointer) and objc_setAssociatedObject doesn't like tagged pointers on (iOS 7.x with 64 bits) I think. If I replace the NSIndexPath (column) category with a plain NSObject, the code works.